Add endpoint to get popular individuals by total votes
Introduced a new DTO and service method to retrieve the most popular individuals based on the total number of votes from movie contributions. Updated the API endpoint to use this new method and added corresponding REST call examples. This provides a more accurate measure of popularity compared to name rating.